Extract FeatureIdsHelper from `Dataframe`
